Refactor utils module
