+++ __tests__/issue.test.ts
@@ -22,7 +22,7 @@ import fs from 'fs';
 import path from 'path';
 import { Application } from 'probot';
 import robot from '../src';
-import { fetchConfigFile, labelExists, loadTemplate } from '../src/libs/utils';
+import { fetchConfigFile, labelExists, loadTemplate } from '../src/libs/ghutils';
 
 const p0 = path.join(__dirname, 'fixtures/issue-comment-created-unassigned.json');
 const unassignedIssueCommentCreated = JSON.parse(fs.readFileSync(p0, 'utf8'));
@@ -53,7 +53,7 @@ jest.mock('../src/libs/repository', () => ({
   addLicenseIfRequired: jest.fn().mockReturnValueOnce(Promise.resolve()),
 }));
 
-jest.mock('../src/libs/utils', () => ({
+jest.mock('../src/libs/ghutils', () => ({
   fetchConfigFile: jest.fn(),
   loadTemplate: jest.fn(),
   labelExists: jest.fn(),
